# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API with Swagger documentation.",
        "title": "UnfallAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/unfallVerunglueckteBundesland": {
            "get": {
                "description": "(46241-0023/46241-0024) Verunglückte: Bundesländer, Jahre, Monate, Ortslage, Schwere der Verletzung",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GENESIS-Online (Die Datenbank des Statistischen Bundesamtes)"
                ],
                "summary": "Verunglückte nach Bundesland abrufen",
                "parameters": [
                    {
                        "enum": [
                            "Baden-Württemberg",
                            "Bayern",
                            "Berlin",
                            "Brandenburg",
                            "Bremen",
                            "Hamburg",
                            "Hessen",
                            "Mecklenburg-Vorpommern",
                            "Niedersachsen",
                            "Nordrhein-Westfalen",
                            "Rheinland-Pfalz",
                            "Saarland",
                            "Sachsen",
                            "Sachsen-Anhalt",
                            "Schleswig-Holstein",
                            "Thüringen"
                        ],
                        "type": "string",
                        "description": "Filter nach Bundesland",
                        "name": "bundesland",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter nach Ortslage (innerorts, außerorts (ohne Autobahnen), auf Autobahnen, Insgesamt)",
                        "name": "ortslage",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "Getötete",
                            "Schwerverletzte",
                            "Leichtverletzte",
                            "Insgesamt"
                        ],
                        "type": "string",
                        "description": "Filter nach Schweregrad",
                        "name": "schweregrad",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter nach Jahr",
                        "name": "jahr",
                        "in": "query"
                    },
                    {
                        "maximum": 12,
                        "minimum": 0,
                        "type": "integer",
                        "description": "Filter nach Monat (1-12, 0 für Ganzjahresdaten)",
                        "name": "monat",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.UnfallVerunglueckteBundesland"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameter type or range",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Database execution or scanning failure",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    }
                }
            }
        },
        "/unfallVerunglueckteBundesland/jahre": {
            "get": {
                "description": "Gibt alle Jahre zurück, für die Daten vorhanden sind.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GENESIS-Online (Die Datenbank des Statistischen Bundesamtes)"
                ],
                "summary": "Verfügbare Jahre abrufen",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.YearsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Database execution or scanning failure",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "data.UnfallVerunglueckteBundesland": {
            "type": "object",
            "properties": {
                "anzahl": {
                    "type": "integer"
                },
                "bundesland": {
                    "type": "string"
                },
                "jahr": {
                    "type": "integer"
                },
                "monat": {
                    "description": "1-12 for months, 0 for full year data",
                    "type": "integer"
                },
                "ortslage": {
                    "type": "string"
                },
                "schweregrad": {
                    "type": "string"
                }
            }
        },
        "handlers.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid parameter format"
                }
            }
        },
        "handlers.YearsResponse": {
            "type": "object",
            "properties": {
                "jahre": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        2022
                    ]
                }
            }
        }
    }
}